update modal design
