+++ packages/datatrak-web/src/features/Tasks/TaskActionsMenu.tsx
@@ -4,13 +4,13 @@
  */
 
 import React, { useState } from 'react';
-import { IconButton, Typography } from '@material-ui/core';
+import { IconButton } from '@material-ui/core';
 import { TaskStatus } from '@tupaia/types';
 import { ActionsMenu } from '@tupaia/ui-components';
 import styled from 'styled-components';
 import { useEditTask } from '../../api';
-import { SmallModal } from '../../components';
-import { SingleTaskResponse, TasksResponse } from '../../types';
+import { CancelTaskModal } from './CancelTaskModal';
+import { TasksResponse } from '../../types';
 
 const MenuButton = styled(IconButton)`
   &.MuiIconButton-root {
@@ -19,40 +19,12 @@ const MenuButton = styled(IconButton)`
   }
 `;
 
-interface ModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-  onCancelTask: () => void;
-  isLoading: boolean;
-}
-
-const CancelTaskModal = ({ isOpen, onClose, onCancelTask, isLoading }: ModalProps) => (
-  <SmallModal
-    open={isOpen}
-    onClose={onClose}
-    title="Cancel task"
-    isLoading={isLoading}
-    primaryButton={{
-      label: 'Cancel task',
-      onClick: onCancelTask,
-    }}
-    secondaryButton={{
-      label: 'Go back',
-      onClick: onClose,
-    }}
-  >
-    <Typography align="center">
-      Are you sure you would like to cancel this task? This cannot be undone.
-    </Typography>
-  </SmallModal>
-);
-
-export const TaskActionsMenu = ({ id, taskStatus }: TasksResponse | SingleTaskResponse) => {
+export const TaskActionsMenu = ({ task }: { task: TasksResponse }) => {
   const [isOpen, setIsOpen] = useState(false);
   const onOpen = () => setIsOpen(true);
   const onClose = () => setIsOpen(false);
 
-  const { mutate, isLoading } = useEditTask(id, onClose);
+  const { mutate, isLoading } = useEditTask(task.id, onClose);
 
   const onCancelTask = () => {
     mutate({ status: TaskStatus.cancelled });
@@ -65,7 +37,10 @@ export const TaskActionsMenu = ({ id, taskStatus }: TasksResponse | SingleTaskRe
     },
   ];
 
-  if (taskStatus === TaskStatus.cancelled || taskStatus === TaskStatus.completed) return null;
+  if (task.taskStatus === TaskStatus.cancelled || task.taskStatus === TaskStatus.completed) {
+    return null;
+  }
+  console.log('ACTIIONS', task);
 
   return (
     <>
@@ -75,6 +50,7 @@ export const TaskActionsMenu = ({ id, taskStatus }: TasksResponse | SingleTaskRe
         onClose={onClose}
         onCancelTask={onCancelTask}
         isLoading={isLoading}
+        task={task}
       />
     </>
   );
